refactor(spi): move the sweep-policy cluster into the spi package
SweepPolicy/StatusTransitionSweepPolicy and their SweepAction/PropositionMark/
MarkReason vocabulary move from projection.memory into com.embabel.dice.spi,
sitting alongside StatusTransitionPolicy. Deciding a marked proposition's
lifecycle fate is a policy seam, the same kind of plug-point as the trust,
authority, conflict, and status policies already in spi. The cluster depends only
on the core proposition model, so spi keeps its 'depends on core only' layering;
the collector in projection.memory now imports these from spi (the correct
feature -> spi direction). Pure move, no behaviour change.

Also relocate StatusTransitionSweepPolicyTest into the spi test package to match
the DecayStatusPolicyTest convention.

Note: moving these public types is a source-incompatible change for any external
importer of the old package (acceptable pre-1.0).
