+++ dice/src/main/kotlin/com/embabel/dice/projection/memory/DefaultCollectorRunner.kt
@@ -26,6 +26,9 @@ import com.embabel.dice.proposition.Proposition
 import com.embabel.dice.proposition.PropositionQuery
 import com.embabel.dice.proposition.PropositionRepository
 import com.embabel.dice.proposition.PropositionStatus
+import com.embabel.dice.spi.PropositionMark
+import com.embabel.dice.spi.SweepAction
+import com.embabel.dice.spi.SweepPolicy
 import org.slf4j.LoggerFactory
 import java.time.Instant
 import java.util.UUID
