+++ dice/src/main/kotlin/com/embabel/dice/spi/MarkReason.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.embabel.dice.projection.memory
+package com.embabel.dice.spi
 
 /**
  * Why a proposition was marked for collection by a [CollectorStrategy].
