+++ dice/src/main/kotlin/com/embabel/dice/spi/SweepAction.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.embabel.dice.projection.memory
+package com.embabel.dice.spi
 
 import com.embabel.dice.proposition.PropositionStatus
