+++ dice/src/test/kotlin/com/embabel/dice/projection/lineage/InMemoryCollectorRecordStoreTest.kt
@@ -15,7 +15,7 @@
  */
 package com.embabel.dice.projection.lineage
 
-import com.embabel.dice.projection.memory.MarkReason
+import com.embabel.dice.spi.MarkReason
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
