+++ dice/src/test/kotlin/com/embabel/dice/projection/memory/CollectorRunnerBuilderTest.kt
@@ -16,6 +16,7 @@
 package com.embabel.dice.projection.memory
 
 import com.embabel.dice.proposition.PropositionRepository
+import com.embabel.dice.spi.StatusTransitionSweepPolicy
 import io.mockk.mockk
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertFalse
