+++ dice/src/test/kotlin/com/embabel/dice/projection/memory/CollectorStrategyTest.kt
@@ -18,6 +18,8 @@ package com.embabel.dice.projection.memory
 import com.embabel.agent.core.ContextId
 import com.embabel.dice.proposition.Proposition
 import com.embabel.dice.proposition.PropositionRepository
+import com.embabel.dice.spi.MarkReason
+import com.embabel.dice.spi.PropositionMark
 import io.mockk.mockk
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertThrows
