+++ dice/src/test/kotlin/com/embabel/dice/projection/memory/DecayCollectorStrategyTest.kt
@@ -18,6 +18,7 @@ package com.embabel.dice.projection.memory
 import com.embabel.agent.core.ContextId
 import com.embabel.dice.proposition.Proposition
 import com.embabel.dice.proposition.PropositionRepository
+import com.embabel.dice.spi.MarkReason
 import io.mockk.mockk
 import io.mockk.verify
 import org.junit.jupiter.api.Assertions.assertEquals
